+++ modules/tinymce/src/core/main/ts/selection/GetSelectionContentImpl.ts
@@ -2,6 +2,7 @@ import { Fun, Optional } from '@ephox/katamari';
 import { SugarElement } from '@ephox/sugar';
 
 import Editor from '../api/Editor';
+import { cleanupBogusElements, cleanupInputNames } from '../content/ContentCleanup';
 import { Content, ContentFormat, GetSelectionContentArgs } from '../content/ContentTypes';
 import { postProcessGetContent, preProcessGetContent } from '../content/PrePostProcess';
 import * as CharType from '../text/CharType';
@@ -26,10 +27,15 @@ const getTextContent = (editor: Editor): string =>
 
     const contextNodeName = getContextNodeName(parentBlockOpt);
 
+    const rangeContentClone = SugarElement.fromDom(rng.cloneContents());
+    cleanupBogusElements(rangeContentClone);
+    cleanupInputNames(rangeContentClone);
+
     const bin = editor.dom.add(body, contextNodeName, {
       'data-mce-bogus': 'all',
       'style': 'overflow: hidden; opacity: 0;'
-    }, rng.cloneContents());
+    }, rangeContentClone.dom);
+
     const text = getInnerText(bin);
 
     // textContent will not strip leading/trailing spaces since it doesn't consider how it'll render