+++ modules/tinymce/src/core/test/ts/browser/delete/BlockBoundaryDeleteTest.ts
@@ -1,5 +1,5 @@
 import { ApproxStructure } from '@ephox/agar';
-import { describe, it } from '@ephox/bedrock-client';
+import { context, describe, it } from '@ephox/bedrock-client';
 import { TinyAssertions, TinyHooks, TinySelections } from '@ephox/wrap-mcagar';
 import { assert } from 'chai';
 
@@ -51,33 +51,6 @@ describe('browser.tinymce.core.delete.BlockBoundaryDeleteTest', () => {
     noopDelete(editor);
     TinyAssertions.assertContent(editor, '<p>a</p>');
     TinyAssertions.assertSelection(editor, [ 0, 0 ], 1, [ 0, 0 ], 1);
-  });
-
-  it('Backspace between blocks with different parents should not merge', () => {
-    const editor = hook.editor();
-    editor.setContent('<p>a</p><div><p>b</p></div>');
-    TinySelections.setCursor(editor, [ 1, 0, 0 ], 0);
-    noopBackspace(editor);
-    TinyAssertions.assertContent(editor, '<p>a</p><div><p>b</p></div>');
-    TinyAssertions.assertSelection(editor, [ 1, 0, 0 ], 0, [ 1, 0, 0 ], 0);
-  });
-
-  it('Delete between blocks with different parents should not merge', () => {
-    const editor = hook.editor();
-    editor.setContent('<p>a</p><div><p>b</p></div>');
-    TinySelections.setCursor(editor, [ 0, 0 ], 1);
-    noopDelete(editor);
-    TinyAssertions.assertContent(editor, '<p>a</p><div><p>b</p></div>');
-    TinyAssertions.assertSelection(editor, [ 0, 0 ], 1, [ 0, 0 ], 1);
-  });
-
-  it('Backspace between textblock and non text block should not merge', () => {
-    const editor = hook.editor();
-    editor.setContent('<p>a</p><ul><li>b</li></ul>');
-    TinySelections.setCursor(editor, [ 1, 0, 0 ], 0);
-    noopBackspace(editor);
-    TinyAssertions.assertContent(editor, '<p>a</p><ul><li>b</li></ul>');
-    TinyAssertions.assertSelection(editor, [ 1, 0, 0 ], 0, [ 1, 0, 0 ], 0);
   });
 
   it('Delete between textblock and non text block should not merge', () => {
@@ -345,5 +318,127 @@ describe('browser.tinymce.core.delete.BlockBoundaryDeleteTest', () => {
       }))
     );
     TinyAssertions.assertSelection(editor, [ 0, 2 ], 1, [ 0, 2 ], 1);
+  });
+
+  context('Nested elements', () => {
+    it('TINY-9230: Delete between two elements where one is nested should merge them', () => {
+      const editor = hook.editor();
+      editor.setContent('<p>a</p><div><p>b</p></div>');
+      TinySelections.setCursor(editor, [ 0, 0 ], 1);
+      doDelete(editor);
+      TinyAssertions.assertContent(editor, '<p>ab</p>');
+      TinyAssertions.assertCursor(editor, [ 0, 0 ], 1);
+    });
+
+    it('TINY-9230: Backspace between two elements where one is nested should merge them', () => {
+      const editor = hook.editor();
+      editor.setContent('<p>a</p><div><p>b</p></div>');
+      TinySelections.setCursor(editor, [ 1, 0 ], 0);
+      doBackspace(editor);
+      TinyAssertions.assertContent(editor, '<p>ab</p>');
+      TinyAssertions.assertCursor(editor, [ 0, 0 ], 1);
+    });
+
+    it('TINY-9230: Delete between two elements where one is nested in a table should not merge them', () => {
+      const editor = hook.editor();
+      editor.setContent('<p>a</p><table><tbody><tr><td><p>b</p></td></tr></tbody></table>');
+      TinySelections.setCursor(editor, [ 0, 0 ], 1);
+      noopDelete(editor);
+    });
+
+    it('TINY-9230: Backspace between two elements where one is nested in a table should not merge them', () => {
+      const editor = hook.editor();
+      editor.setContent('<p>a</p><table><tbody><tr><td><p>b</p></td></tr></tbody></table>');
+      TinySelections.setCursor(editor, [ 1, 0, 0, 0 ], 0);
+      noopBackspace(editor);
+    });
+
+    it('TINY-9230: Delete between two cells should be a noop', () => {
+      const editor = hook.editor();
+      editor.setContent('<table><tbody><tr><td><p>b</p></td><td><p>b</p></tr></tbody></table>');
+      TinySelections.setCursor(editor, [ 0, 0, 0, 0, 0, 0 ], 1);
+      noopDelete(editor);
+    });
+
+    it('TINY-9230: Backspace between two cells should be a noop', () => {
+      const editor = hook.editor();
+      editor.setContent('<table><tbody><tr><td><p>b</p></td><td><p>b</p></tr></tbody></table>');
+      TinySelections.setCursor(editor, [ 0, 0, 0, 1, 0, 0 ], 0);
+      noopBackspace(editor);
+    });
+
+    it('TINY-9230: Delete between two editing hosts should be a noop', () => {
+      const editor = hook.editor();
+      editor.setContent('<div contenteditable="false"><div contenteditable="true">a</div></div><div contenteditable="false"><div contenteditable="true">b</div></div>');
+      TinySelections.setCursor(editor, [ 1, 0, 0 ], 1); // Start index off by one to compensate for fake caret paragraph
+      noopDelete(editor);
+    });
+
+    it('TINY-9230: Backspace between two editing hosts should be a noop', () => {
+      const editor = hook.editor();
+      editor.setContent('<div contenteditable="false"><div contenteditable="true">a</div></div><div contenteditable="false"><div contenteditable="true">b</div></div>');
+      TinySelections.setCursor(editor, [ 2, 0, 0 ], 0); // Start index off by one to compensate for fake caret paragraph
+      noopBackspace(editor);
+    });
+  });
+
+  context('Transparent elements', () => {
+    it('TINY-9230: Delete between two transparent blocks should merge them', () => {
+      const editor = hook.editor();
+      editor.setContent('<a href="#1"><p>a</p></a><a href="#2"><p>b</p></a>');
+      TinySelections.setCursor(editor, [ 0, 0, 0 ], 1);
+      doDelete(editor);
+      TinyAssertions.assertContentStructure(editor,
+        ApproxStructure.build((s, str, _arr) => s.element('body', {
+          children: [
+            s.element('a', {
+              attrs: {
+                'href': str.is('#1'),
+                'data-mce-block': str.is('true')
+              },
+              children: [
+                s.element('p', {
+                  children: [
+                    s.text(str.is('a')),
+                    s.text(str.is('b'))
+                  ]
+                })
+              ]
+            })
+          ]
+        }))
+      );
+      TinyAssertions.assertContent(editor, '<a href="#1"><p>ab</p></a>');
+      TinyAssertions.assertCursor(editor, [ 0, 0, 0 ], 1);
+    });
+
+    it('TINY-9230: Backspace between two transparent blocks should merge', () => {
+      const editor = hook.editor();
+      editor.setContent('<a href="#1"><p>a</p></a><a href="#2"><p>b</p></a>');
+      TinySelections.setCursor(editor, [ 1, 0, 0 ], 0);
+      doBackspace(editor);
+      TinyAssertions.assertContentStructure(editor,
+        ApproxStructure.build((s, str, _arr) => s.element('body', {
+          children: [
+            s.element('a', {
+              attrs: {
+                'href': str.is('#1'),
+                'data-mce-block': str.is('true')
+              },
+              children: [
+                s.element('p', {
+                  children: [
+                    s.text(str.is('a')),
+                    s.text(str.is('b'))
+                  ]
+                })
+              ]
+            })
+          ]
+        }))
+      );
+      TinyAssertions.assertContent(editor, '<a href="#1"><p>ab</p></a>');
+      TinyAssertions.assertCursor(editor, [ 0, 0, 0 ], 1);
+    });
   });
 });